+++ src/core/base.jl
@@ -90,7 +90,10 @@ function GenericPowerModel(data::Dict{String,Any}, T::DataType; ext = Dict{Strin
 end
 
 
-### Helper functions for ignoring multinetwork support
+### Helper functions for working with multinetworks
+ismultinetwork(pm::GenericPowerModel) = (length(pm.ref[:nw]) > 1)
+nws(pm::GenericPowerModel) = keys(pm.ref[:nw])
+
 ids(pm::GenericPowerModel, key::Symbol) = ids(pm, pm.cnw, key)
 ids(pm::GenericPowerModel, n::Int, key::Symbol) = keys(pm.ref[:nw][n][key])
 
@@ -158,8 +161,8 @@ function build_generic_model(data::Dict{String,Any}, model_constructor, post_met
     # NOTE, this model constructor will build the ref dict using the latest info from the data
     pm = model_constructor(data; kwargs...)
 
-    if !multinetwork && data["multinetwork"]
-        warn(LOGGER, "building a single network model with multinetwork data, only network ($(pm.cnw)) will be used.")
+    if !multinetwork && ismultinetwork(pm)
+        error(LOGGER, "attempted to build a single-network model with multi-network data")
     end
 
     post_method(pm)